Added a configurable potion list expansion.
This is based off of a little tutorial I just finished on reflection,
and will expand the list to any number the user chooses. This is of dire
need; previously, the list only goes to ~30.

Also, I figured out why the entity isn't appearing. We needed to
register our entity with a render class. I have a shell render class,
which I will add into later.