Added some cool things to potion effects.
Tweaked the colors of the potion particles, and added the following to
them -

The combustion potion effect, when applied to a creeper, will charge it
instead of causing it to explode. The combustion effect will also not
cause the player to explode if they are in creative mode.

The angel effect will cause great amounts of damage to the undead.