Added two more potions.
We have the Potion of Cryogenics, a potion that reduces movement speed
of an entity down to next to zero. This includes a Long variant. Brewed
from a snowball (for now.)

Along with this, the Potion of Stability, which reduces the amount of
knockback an entity takes from damage. Long variants. Brewed with the
Obsidian Tear, crafted from a Ghast Tear and obsidian, or found as a
rare drop from Ghasts.

I'll add in the potion icons later, along with the new fluid textures.