Added lore to potions.
I don't know, here's my take on lore for the potions. I'm liking it,
what do you think?

Notes -

I'm fairly sure I'll be changing a couple of the descriptions of the
potions. A couple of them are vague and don't look to professional.
Otherwise, it's like the vanilla system for the most part.