Add the Pure Tear (desc).
The Pure Tear is the counterpart to the Plagued Tear. Ghast do not drop
this tear. Instead, villagers can trade for the Pure Tear now. When a
player enters the Nether (this functionality will be added soon), any
Pure tears they have will be transformed into Plagued Tears, and the
same with Plagued Tears when entering the overworld.