Made ItemTear a MetaItem.
Also fixed it's onUpdate, and added a version that can be called by the
Brewery, which it's now doing in the pre work phase in order to prevent
users from circumventing the nether rules with automation.

FluidPotion now set's it's unlocalized name to that of the inputted
item. No idea if this will work or not, but it'll sure help if it does.