Lots of fixes. It starts up now.
Attempted to make packet system for potion entity, but was stopped by
fatal bug in Forge, so that code has been commented out.

Put a permanent solution for the brewering rendering issues. Got rid of
the proxies and made the client side just a plugin.

Fixed a localization and the name of obsidiantear. (I had capped the T)

FluidPotion only get's it's color on client side, since that method is
client only. Shouldn't cause any problems as far as I can imagine.

Fixed potion item not drinking or throwing properly.

Added methods to PotionRegistry that I forgot.

Changes to brewery and it's render require updating Core.