A couple of fixes.
Fixed a couple potions having descriptions that weren't supposed to,
fixed a couple of localizations not working, added in something I forgot
to do in MetaItemPotion, and added the missing splash bottle texture. I
figure it isn't worth separately registering the icons for a generic sub
item just to use the vanilla icon.