Moved a class and cleaned the config a bit.
The config now works with individual potions because we use the display
name, which is now based on strength and duration. Removed a basically
redundant comments that went with each of the config entries, and moved
the creative tab entry to the 'general' category.

Also moved the creative tab class to the utils package.