Made the witch villager! (desc)
It was surprisingly easy to do, just two lines of code. I did attempt to
create the witch hut-type building for village generation, but that
failed miserably, miserably enough where I ripped out all of that code
before committing.

I'll get to creating the village building whenever I feel like I can
take that on.

Updated the properties file for version 1.2.0, and changed the name of a
class.