Added a non-functional Sprayer block.
I created all necessary classes, including the factory, tile entity, and
such. All slots are in place, the Gui is finished. No potions are
sprayed, though. We're going to probably need to inherit IRedstoneTile
(just found that, don't know how to use it), and have it spray one
potion per redstone signal, sort of like a dispenser. We may also need a
new MetaTile extension that can handle spraying particles in a cloud and
inflicting a potion effect from the item. For now, the block is simply a
MetaTile.