Re-added ore dictionary Kegs.
Okay, so I can't have them be only present when the metals are present,
because of the fact that icon registry for the models runs off of
metadata, and so does the tooltip system. Therefore, I have included all
of the Kegs in the creative menu (icons not present because I'm waiting
for Technic's Kegs model to skin) but am also including a tooltip line
if the metal returns null - 'Uncraftable; metal not found'

I think this should do it for Kegs until the new model comes in. . .