Big Potions & Vials & Warlocks, oh my!
Pretty big commit. *deep breath*

- Added support for custom potion textures and fluid capacities.
- Got most of the workings done for adding delay to the Sprayer, now
just for getting the Sprayer to update the delay when it changes.
- Added Warlocks (don't worry, not a new mob. :p) It's a counterpart to
the villager witch, who only sells ingredients.
- Changed the villager witch to only sell potions.
- Added multiple creative tabs.
- Added Big Potions (2000 mB), and Vials (250 mB).
- Fixed some of the vanilla potions' III-level potions not working
properly.
- Am starting to drift away from the event class to handle tooltips, got
a couple crashes from that for no reason in particular.
- Moved some of the items in the 'ingredients' meta item to a
miscellaneous item.
- Removed a potion that doesn't scale with the potion levels.
- API changes, to support new things! This will need to be changed, I
don't like the way I have extra methods that have to be called to create
a potion.
- Changed the creative tab background a little. (I may want to revert
this back to vanilla, it's causing some irritating visual glitches with
unopened tab icons.)
- Changed shape of sprayer GUI back to vanilla.