Sprayer Gui works
Had to mash some code into ParticleHandler to send packets for the GUI.
This is horrible and ugly! It works for now, but a real packet handler
is needed.

Also worked the Tile over to run better. Delay is in seconds, with 1 as
the lowest and 30 the highest. Idle is reset whenever redstone state
changes, so it should seem very responsive even when it's on long
timers.